+++ lib/i18n-js.rb
@@ -23,29 +23,25 @@ module I18nJS
             "you must set either `config_file` or `config`"
     end
 
-    load_plugins!
-
     config = Glob::SymbolizeKeys.call(config || load_config_file(config_file))
 
+    load_plugins!
+    initialize_plugins!(config: config)
     Schema.validate!(config)
+
     exported_files = []
 
-    config[:translations].each do |group|
-      exported_files += export_group(group, config)
-    end
-
-    plugins.each do |plugin|
-      plugin.after_export(files: exported_files.dup, config: config)
-    end
+    config[:translations].each {|group| exported_files += export_group(group) }
+    plugins.each {|plugin| plugin.after_export(files: exported_files.dup) }
 
     exported_files
   end
 
-  def self.export_group(group, config)
+  def self.export_group(group)
     filtered_translations = Glob.filter(translations, group[:patterns])
     filtered_translations =
       plugins.reduce(filtered_translations) do |buffer, plugin|
-        plugin.transform(translations: buffer, config: config)
+        plugin.transform(translations: buffer)
       end
 
     output_file_path = File.expand_path(group[:file])